refactor(packages): create token and productcategory packages in model package
