--- worldoftoys/src/main/java/com/kopchak/worldoftoys/model/AuthTokenType.java
@@ -1,5 +1,0 @@
-package com.kopchak.worldoftoys.model;
-
-public enum AuthTokenType {
-    BEARER
-}
