--- worldoftoys/src/main/java/com/kopchak/worldoftoys/model/GenderCategory.java
@@ -1,14 +1,0 @@
-package com.kopchak.worldoftoys.model;
-
-import jakarta.persistence.Entity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-public class GenderCategory extends ProductCategories{
-
-}
