+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/User.java
@@ -1,5 +1,7 @@
 package com.kopchak.worldoftoys.model;
 
+import com.kopchak.worldoftoys.model.token.AuthenticationToken;
+import com.kopchak.worldoftoys.model.token.ConfirmationToken;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -8,7 +10,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.lang.NonNull;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
