+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/productcategory/OriginCategory.java
@@ -1,0 +1,15 @@
+package com.kopchak.worldoftoys.model.productcategory;
+
+import com.kopchak.worldoftoys.model.productcategory.ProductCategories;
+import jakarta.persistence.Entity;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+public class OriginCategory extends ProductCategories {
+
+}
