+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/token/AuthTokenType.java
@@ -1,0 +1,5 @@
+package com.kopchak.worldoftoys.model.token;
+
+public enum AuthTokenType {
+    BEARER
+}
