+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/model/token/AuthenticationToken.java
@@ -1,5 +1,6 @@
-package com.kopchak.worldoftoys.model;
+package com.kopchak.worldoftoys.model.token;
 
+import com.kopchak.worldoftoys.model.User;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,7 +9,7 @@ import org.springframework.lang.NonNull;
 @Data
 @NoArgsConstructor
 @Entity
-public class AuthenticationToken extends Token{
+public class AuthenticationToken extends Token {
     @Enumerated(EnumType.STRING)
     @NonNull
     private AuthTokenType tokenType;
