+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/service/ConfirmationTokenService.java
@@ -2,7 +2,7 @@ package com.kopchak.worldoftoys.service;
 
 import com.kopchak.worldoftoys.dto.ConfirmTokenDto;
 import com.kopchak.worldoftoys.dto.PasswordResetDto;
-import com.kopchak.worldoftoys.model.ConfirmTokenType;
+import com.kopchak.worldoftoys.model.token.ConfirmTokenType;
 
 public interface ConfirmationTokenService {
     ConfirmTokenDto createConfirmToken(String username, ConfirmTokenType tokenType);
