+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/service/impl/ConfirmationTokenServiceImpl.java
@@ -5,8 +5,8 @@ import com.kopchak.worldoftoys.dto.PasswordResetDto;
 import com.kopchak.worldoftoys.exception.ConfirmationTokenExpiredException;
 import com.kopchak.worldoftoys.exception.NewPasswordMatchesOldException;
 import com.kopchak.worldoftoys.exception.UserNotFoundException;
-import com.kopchak.worldoftoys.model.ConfirmTokenType;
-import com.kopchak.worldoftoys.model.ConfirmationToken;
+import com.kopchak.worldoftoys.model.token.ConfirmTokenType;
+import com.kopchak.worldoftoys.model.token.ConfirmationToken;
 import com.kopchak.worldoftoys.model.User;
 import com.kopchak.worldoftoys.repository.ConfirmationTokenRepository;
 import com.kopchak.worldoftoys.repository.UserRepository;
